+++ Audit.NET.sln
@@ -152,6 +152,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Audit.Kafka.UnitTest", "tes
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Audit.MongoDB.UnitTest", "test\Audit.MongoDB.UnitTest\Audit.MongoDB.UnitTest.csproj", "{FDA71427-459F-42CE-9607-CDD6EBE1BEF8}"
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Audit.MySql.UnitTest", "test\Audit.MySql.UnitTest\Audit.MySql.UnitTest.csproj", "{D81558A2-EC45-4526-88E2-F385D42E79E6}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -656,6 +658,14 @@ Global
 		{FDA71427-459F-42CE-9607-CDD6EBE1BEF8}.Release|Any CPU.Build.0 = Release|Any CPU
 		{FDA71427-459F-42CE-9607-CDD6EBE1BEF8}.Release|x64.ActiveCfg = Release|Any CPU
 		{FDA71427-459F-42CE-9607-CDD6EBE1BEF8}.Release|x64.Build.0 = Release|Any CPU
+		{D81558A2-EC45-4526-88E2-F385D42E79E6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D81558A2-EC45-4526-88E2-F385D42E79E6}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D81558A2-EC45-4526-88E2-F385D42E79E6}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{D81558A2-EC45-4526-88E2-F385D42E79E6}.Debug|x64.Build.0 = Debug|Any CPU
+		{D81558A2-EC45-4526-88E2-F385D42E79E6}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D81558A2-EC45-4526-88E2-F385D42E79E6}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D81558A2-EC45-4526-88E2-F385D42E79E6}.Release|x64.ActiveCfg = Release|Any CPU
+		{D81558A2-EC45-4526-88E2-F385D42E79E6}.Release|x64.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -725,6 +735,7 @@ Global
 		{04D5CCB9-E095-40D3-9986-C1E41D8F638F} = {A54B4BB6-3439-432B-AFD9-FE62D6528D42}
 		{F812D4F7-322D-48D0-A275-1AEC6978ED23} = {A54B4BB6-3439-432B-AFD9-FE62D6528D42}
 		{FDA71427-459F-42CE-9607-CDD6EBE1BEF8} = {A54B4BB6-3439-432B-AFD9-FE62D6528D42}
+		{D81558A2-EC45-4526-88E2-F385D42E79E6} = {A54B4BB6-3439-432B-AFD9-FE62D6528D42}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {B1989475-43D6-4AA5-9717-6DBF734B0C6E}
