+++ test/Audit.IntegrationTest/Audit.IntegrationTest.csproj
@@ -24,7 +24,6 @@
     <ProjectReference Include="..\..\src\Audit.NET.Udp\Audit.NET.Udp.csproj" />
     <ProjectReference Include="..\..\src\Audit.NET.SqlServer\Audit.NET.SqlServer.csproj" />
     <ProjectReference Include="..\..\src\Audit.NET.AzureStorage\Audit.NET.AzureStorage.csproj" />
-    <ProjectReference Include="..\..\src\Audit.NET.MySql\Audit.NET.MySql.csproj" />
     <ProjectReference Include="..\..\src\Audit.Mvc\Audit.Mvc.csproj" />
     <ProjectReference Include="..\..\src\Audit.WebApi\Audit.WebApi.csproj" />
   </ItemGroup>
